Add world rendering functionality
The GameApplication class now is initialized with a world object, which
is rendered to the screen each frame. This is done using a new constant
denoting the size of each tile. Currently, as no camera functionality
has been implemented, the display size is determined by the world size
and tile size.