Add new bounds tile and bounds checking for world
A new bounds tile represents a tile that is out of bounds of a game
world. Now, when a tile is queried from a world object, if the
coordinates are out of bounds, the bounds tile is returned. This will
make it a lot easier to render the game world once camera movement is
implemented. It will help prevent issues once the camera moves outside
the game world, as the bounds tile will be rendered instead of a normal
tile.