Offload world rendering to camera class
The camera class now handles the rendering of the world. This is done by
passing a PyGame surface to the render method in a camera object. The
camera renders the world to the surface taking into account the current
camera position, and the size of the window.

This does assume the window size is equal to the camera size multiplied
by the tile size, which raises issues if this is not the case. If time,
this will be improved in a future commit.