Add functionality to create and render entities
The ImageEntity class represents an entity with x and y world
coordinates with an graphical image associated with it, in the form
of a PyGame surface object.

The camera class has been updated with a method to render an image
entity relative to the game world, centering the surface on the
corresponding tile.

In order for a sprite to fit within a tile, it is important that the
corresponding image is less than or equal to the tile size. The game
tile size and window settings have been adjusted to the minimum size
needed to display the included images within a tile, without needing to
scale the resolution of each image.

The GameApplication class has been adjusted to reflect these changes.
Some example entities have been added, along with a player robot entity.
The players position can be controlled with the arrow keys, and the
camera is now centered on the player. This example code is subject to
change and refactoring.

In the future, better player movement will be implemented, along with
tile collision checking.