Add some basic code to repeat player movement
You can now hold a key down and player will continue moving. This was
done using the pygame.time.set_timer function to schedule a custom
player move event every time a movement key is pressed. This event is
stopped when both the player x velocity and y velocity are 0. This
creates an issue where the player stops moving if both left and right
(or up and down) are pressed, then one is released. This is subject
to improvement.