Implement smoother movement and collision checking
Improved the code for player movement to create smoother input. This
was done by extending the entity class to add additional functionality
to a dedicated player class. Collision has also been implemented, by
passing a world object to the Player.move method to check if the
destination tile is collidable.