Add code to make monsters move around
Monsters now move around second. They move 1 tile either up down right
or left, and will not move at all if the random new tile is a
collidable one, or if it will move them on top of a coin. This is a very
crude AI setup, but for the scope of this project it is acceptable.
Subject to optimization and improvement in the future.